feature: added basic panel menu
