+++ src/components/Panel/PanelMenu.tsx
@@ -1,0 +1,31 @@
+import { Add, Help } from "@mui/icons-material";
+import {
+  Divider,
+  ListItemIcon,
+  ListItemText,
+  MenuItem,
+  MenuList,
+  Paper,
+} from "@mui/material";
+
+export default function PanelMenu() {
+  return (
+    <Paper sx={{ width: 320, maxWidth: "100%" }}>
+      <MenuList dense={true}>
+        <MenuItem>
+          <ListItemIcon>
+            <Add fontSize="small" />
+          </ListItemIcon>
+          <ListItemText>Add</ListItemText>
+        </MenuItem>
+        <Divider />
+        <MenuItem>
+          <ListItemIcon>
+            <Help fontSize="small" />
+          </ListItemIcon>
+          <ListItemText>Hint</ListItemText>
+        </MenuItem>
+      </MenuList>
+    </Paper>
+  );
+}
